提交代码

+++ txlcn-tc/src/main/java/com/codingapi/txlcn/tc/corelog/txc/TxcLogHelper.java
@@ -36,27 +36,27 @@ import java.util.List;
 @Slf4j
 public class TxcLogHelper implements LogHelper {
 
-    private final H2DbHelper h2DbHelper;
+    @Autowired(required = false)
+    private H2DbHelper h2DbHelper;
 
-    @Autowired
-    public TxcLogHelper(H2DbHelper h2DbHelper) {
-        this.h2DbHelper = h2DbHelper;
-    }
 
     /**
      * TXC撤销日志存储表准备
      */
     @Override
     public void init() {
-        h2DbHelper.update("CREATE TABLE IF NOT EXISTS TXC_UNDO_LOG (" +
-                "ID BIGINT NOT NULL AUTO_INCREMENT, " +
-                "UNIT_ID VARCHAR(32) NOT NULL," +
-                "GROUP_ID VARCHAR(64) NOT NULL," +
-                "SQL_TYPE INT NOT NULL," +
-                "ROLLBACK_INFO BLOB NOT NULL," +
-                "CREATE_TIME CHAR(23) NOT NULL, " +
-                "PRIMARY KEY(ID) )");
-        log.info("Txc log table finish (H2 DATABASE)");
+        if(h2DbHelper!=null) {
+            h2DbHelper.update("CREATE TABLE IF NOT EXISTS TXC_UNDO_LOG (" +
+                    "ID BIGINT NOT NULL AUTO_INCREMENT, " +
+                    "UNIT_ID VARCHAR(32) NOT NULL," +
+                    "GROUP_ID VARCHAR(64) NOT NULL," +
+                    "SQL_TYPE INT NOT NULL," +
+                    "ROLLBACK_INFO BLOB NOT NULL," +
+                    "CREATE_TIME CHAR(23) NOT NULL, " +
+                    "PRIMARY KEY(ID) )");
+            log.info("Txc log table finish (H2 DATABASE)");
+        }
+        log.info("TxcLogHelper.init 执行完毕! " + h2DbHelper==null?" h2DbHelper is null!":"");
     }
 
     /**
@@ -66,6 +66,8 @@ public class TxcLogHelper implements LogHelper {
      * @throws SQLException 数据库操作失败
      */
     public void saveUndoLog(UndoLogDO undoLogDO) throws SQLException {
+        if(h2DbHelper==null)
+            throw new RuntimeException("tx-lcn.logger必须进行配置!");
         String sql = "INSERT INTO TXC_UNDO_LOG (UNIT_ID,GROUP_ID,SQL_TYPE,ROLLBACK_INFO,CREATE_TIME) VALUES(?,?,?,?,?)";
         h2DbHelper.queryRunner().update(sql, undoLogDO.getUnitId(), undoLogDO.getGroupId(), undoLogDO.getSqlType(),
                 undoLogDO.getRollbackInfo(), undoLogDO.getCreateTime());
@@ -80,6 +82,8 @@ public class TxcLogHelper implements LogHelper {
      * @throws SQLException 数据库操作失败
      */
     public List<UndoLogDO> getUndoLogByGroupAndUnitId(String groupId, String unitId) throws SQLException {
+        if(h2DbHelper==null)
+            throw new RuntimeException("tx-lcn.logger必须进行配置!");
         String sql = "SELECT * FROM TXC_UNDO_LOG WHERE GROUP_ID = ? and UNIT_ID = ?";
         return h2DbHelper.queryRunner().query(sql, rs -> {
             List<UndoLogDO> undoLogDOList = new ArrayList<>();
@@ -104,6 +108,8 @@ public class TxcLogHelper implements LogHelper {
      * @throws SQLException 数据库操作失败
      */
     public void deleteUndoLog(String groupId, String unitId) throws SQLException {
+        if(h2DbHelper==null)
+            throw new RuntimeException("tx-lcn.logger必须进行配置!");
         String sql = "DELETE FROM TXC_UNDO_LOG WHERE GROUP_ID=? AND UNIT_ID=?";
         h2DbHelper.queryRunner().update(sql, groupId, unitId);
     }
